Se modifican los DAO y los Business object parta que queden como los de Auto y se arman las panatallas usando MiPanel

+++ TallerMecanico/src/servicios/UsuarioBusinessObjectImpl.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import dao.UsuarioDAO;
 import entidades.Usuario;
+import excepciones.CamposVaciosException;
 import excepciones.ExistingUserException;
 import excepciones.NoIdObtainedException;
 import excepciones.NonExistingUserException;
@@ -19,36 +20,68 @@ public class UsuarioBusinessObjectImpl implements UsuarioBusinessObject {
 	}
 
 	@Override
-	public Usuario selectUsuario(String usuario, String contraseña)
-			throws TallerMecanicoException, NonExistingUserException {
-		return usuarioDAO.selectUsuario(usuario, contraseña);
+	public Usuario selectUsuario(String usuario)
+			throws TallerMecanicoException, NonExistingUserException, CamposVaciosException {
+		if (!usuario.equals("")) {
+			if (usuarioDAO.selectUsuario(usuario) == null) {
+				throw new NonExistingUserException("Usuario y/o password incorrectos");
+			} else {
+				return usuarioDAO.selectUsuario(usuario);
+			}
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
 	public void insertUsuario(Usuario usuario)
-			throws TallerMecanicoException, ExistingUserException, NoIdObtainedException {
-		
-		usuarioDAO.insertUsuario(usuario);
+			throws TallerMecanicoException, ExistingUserException, NoIdObtainedException, CamposVaciosException {
+		if (validarUsuario(usuario)) {
+			if (usuarioDAO.selectUsuario(usuario.getUsuario()) == null) {
+				usuarioDAO.insertUsuario(usuario);
+			} else {
+				throw new ExistingUserException("El usuario que intenta dar de alta ya existe");
+			}
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
-	public void updateUsuario(Usuario usuario) throws TallerMecanicoException, NonExistingUserException {
-		usuarioDAO.updateUsuario(usuario);
+	public void updateUsuario(Usuario usuario)
+			throws TallerMecanicoException, NonExistingUserException, CamposVaciosException {
+		if (validarUsuario(usuario)) {
+			usuarioDAO.updateUsuario(usuario);
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
-	public void deleteUsuario(String usuario) throws TallerMecanicoException, NonExistingUserException {
-		usuarioDAO.deleteUsuario(usuario);
-	}
-
-	@Override
-	public boolean login(String usuario, String contraseña) throws TallerMecanicoException, NonExistingUserException {
-		usuarioDAO.login(usuario, contraseña);
-		return false;
+	public void deleteUsuario(String usuario)
+			throws TallerMecanicoException, NonExistingUserException, CamposVaciosException {
+		if (!usuario.equals("")) {
+			if (usuarioDAO.selectUsuario(usuario) == null) {
+				throw new NonExistingUserException("El usuario que intenta eliminar no existe");
+			} else {
+				usuarioDAO.deleteUsuario(usuario);
+			}
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
 	public List<Usuario> selectAll() throws TallerMecanicoException {
 		return usuarioDAO.selectAll();
+	}
+
+	private boolean validarUsuario(Usuario usuario) {
+		if (!usuario.getNombre().equals("") && !usuario.getApellido().equals("") && !usuario.getMail().equals("")
+				&& !usuario.getUsuario().equals("") && !usuario.getPassword().equals("")) {
+			return true;
+		} else {
+			return false;
+		}
 	}
 }
